Fix various race conditions in the zip downloading.
If IDEA starts and the plug-in needs to download a new Kobalt distribution, the previous code was trying to access it before it was downloaded.

This change modified `KobaltApplicationComponent.kobaltJar` to be a Future<Path> which will only return that file once it's been downloaded and installed.

Also, killing and restarting servers should be much more robust now.